pyproject.toml
[project]
name = "gnomecast"
dynamic = ["version"]
description = "A native Linux GUI for Chromecasting local files."
readme = "README.md"
license = { text = "GPL-3.0-or-later" }
authors = [
  { name = "Derek Anderson", email = "public@kered.org" }
]
urls = { Homepage = "https://github.com/keredson/gnomecast" }
requires-python = ">=3.10"
dependencies = [
    "bottle~=0.13.3",
    "dbus-python>=1.4.0",
    "html5lib~=1.1",
    "paste~=3.10.1",
    "pycaption~=2.2.16",
    "pychromecast~=13.1.0",
    "pygobject~=3.52.3",
]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: End Users/Desktop",
  "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
]

[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "gnomecast/version.py"

[project.gui-scripts]
gnomecast = "gnomecast:main"

[dependency-groups]
dev = [
    "ruff>=0.11.13",
]
